requirements.txt
streamlit
torch
torchvision
torchaudio
matplotlib
numpy
pillow
